Adds a new layout for pages called page that introduces an 8 column left area for primary content and a 4 column right area for sidebar content via yield_content.
Previously all pages redundantly declared their own Jumbotron. The jumbotron markup has been moved into a partial that accepts parameters and the jumbotron is now called thru yield_content. This allows a template to be as custom as it would like such as the homepage, or standardized like all other pages thru the jumbotron partial & parameters. As Jumbotrons were commmon to all pages this yield_content call is in the global layout.erb. Not all pages (such as the homepage) will need sidebar content so page is a nested layout. The nested layout page has a yield_content for sidebar content so that additional content or in place of the sidebar default social media content from Facebook & Twitter can be used.

A partial was added for the Facebook specific JavaScript SDK calls that need to be inserted after the body tag and the calling js function was added to the new partial sidebar.erb. Twitter's respective content is also added to sidebar.erb.

The pages about-us, team, new-members were all updated to use the new page nested layout, the homepage was updated to use the new layout