Changes the team data from json format to yml
In order to make onboarding the widest range of talent into the brigade as easy as possible, the experience as to be simple. The JSON format looked intimidating for new or non technical members. I want to try using YML to see if its easier for regular people to add themselves in a browser based pull-request workflow. If it's terrible we can switch it back.