Creates a partners page, addressing Issue #24
Essentially a clone of the team page, adds a data/partners.yml with the data fields requested in the issue. Without an sample data the layout in the page is just a guess wrapping the partner's URL around the name of the organization, using a bootrap heading small class for the description of the organization and just printing out the organization's motivation for working with Code for Charlotte. All the panel bodies are set to be column 6 in one giant row so they'll just auto wrap around if there is an even or odd number of partners and bypass the need to do row/column counting.