Replaces Google Universal Analytics code in partials/_javascripts.erb with the middleman extension middleman-google-analytics. Fixes issue #20
The advantage of making the change is that the extension has a setting for not displaying the analytics code during development so while I'm working on features I'm not driving up pageviews. The change requires adding the extension to the Gemfile, running 'bundle install' so Gemfile.lock gets updated, and adding the relevant call to our layout.erb after the footer. Easy Peasy