Adds a helper function that extends padrino's link_to to check if the current page URL matches the URL in the navigation and add the twitter active class
This was such an easy fix that as vexed me. I knew there had to be some way to do this but I hadn't really understand the right way within the middleman process to figure where to do this myself.
I go lucky and found https://github.com/middleman/middleman/issues/303#issuecomment-4362124