Rearranges the navigation and uses the previously added nav_link_to helper so that pages in the navigation will properly use the bootstrap active class
Collapsed a number of related but seperate pages into a dropdown menu called about. The renamed about-us to about that has information primarly on the when & why the Code for Charlotte brigade was formed and it's mission, the captains/team member listing was a small panel on the about-us page, and the recenly added partners page. All of these are topics somone would want to know 'about'

Since another target is to increase the bread and depth of new members, I moved the instructions for how new members 'new-members.html' from the footer up to the navigation.

Finally 'Home' & 'New Members' make use of the helper nav_link_to helper so that when you click on those pages and only those pages will those list items be highlighted by bootstrap to indicate the current page. I did use the shortcut helper on the other current navigation links because they're all links away from the site.