Fixes a problem where email addresses that were null were causing the page build to fail. Also corrects layout errors caused when no email address was present
I wasn't checking that email.blank? before handing the email to the md5 of gravatar generation. So I had to check for both nil and empty strings. Once I had some users who preferred not to add an email address and rolled through the whole array I found that layouts were janky. Specifically 6 column entries with no image didn't take the full 6 columns so they'd collapse and the effect was a jumble of names than nice columns of team members. Now the team members are sorted by email address first, so they are printed in nice 6 column fields then fill in with everybody else. If that looks terrible as more team members add themselves we'll fix it then.