Adds a contact page
Using the Google Drive API we can use jquery to make an ajax request to post the user data to a Google spreadsheet. However the response from Google fails because a COR request, but the post still works. Knowing this a I use the error message of failure to update a div thanking the user for using the form. It's really horrible but it sorta works.