Updates mailing list forms, navigation, and file paths
Code for Charlotte has their own mailchimp account now. The primary mailing list is the only mailing list for the organization so the home page sign up form has been updated to reflect it. As CfC has their own mailchimp account the need for a separate list for the City Mark has gone away. So that form has been removed from the page describing the project.

Secondly CfC has the ability to receive donations to the umbrella CfA organization and so soon we'll have a Donate page instruting companies and organizations of those precedures. Ergo having both 'Donate' and 'Contribute' in the navigation is confusing, I moved the website project from toolkit to projects and renamed the navigation element.